+++ app/javascript/components/Sessions/SessionsForm/index.ts
@@ -1,0 +1,1 @@
+export { SessionsForm } from "./SessionsForm";
